Телеграм бот

+++ tg_bot/src/main/java/com/skillbox/cryptobot/service/CryptoCurrencyService.java
@@ -5,19 +5,20 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.concurrent.atomic.AtomicReference;
 
 @Service
 @Slf4j
 public class CryptoCurrencyService {
-    private final AtomicReference<Double> price = new AtomicReference<>();
+    private final AtomicReference<BigDecimal> price = new AtomicReference<>();
     private final BinanceClient client;
 
     public CryptoCurrencyService(BinanceClient client) {
         this.client = client;
     }
 
-    public double getBitcoinPrice() throws IOException {
+    public BigDecimal getBitcoinPrice() throws IOException {
         if (price.get() == null) {
             price.set(client.getBitcoinPrice());
         }
